feature: add lanes

+++ Accropoly/Assets/0_Scripts/Utility/SaveSystem/Serialization/Custom/Deserializer.cs
@@ -65,7 +65,7 @@ public partial class Deserializer
     }
     public Waypoint Deserialize(Waypoint data)
     {
-        data.pos = Deserialize(new float2());
+        data.pos = Deserialize(new int2());
         return data;
     }
     public MapData Deserialize(MapData data)
